+++ base/AirModel.ts
@@ -18,6 +18,7 @@ import { getFormConfig, getFormConfigList } from '../decorator/FormField'
 import { getSearchConfigList } from '../decorator/SearchField'
 import { getTableConfigList } from '../decorator/TableField'
 import { IJson } from '../interface/IJson'
+import { ClassConstructor } from '../type/ClassConstructor'
 
 /**
  * # 模型超类
@@ -29,7 +30,7 @@ export class AirModel {
    * 会自动进行数据别名转换
    * @param json `JSON`
    */
-  static fromJson<T extends AirModel>(this: new () => T, json: IJson = {}): T {
+  static fromJson<T extends AirModel>(this: ClassConstructor<T>, json: IJson = {}): T {
     const instance: T = (Object.assign(new this()) as T)
     return AirModel.parse<T>(instance, json)
   }
@@ -75,7 +76,7 @@ export class AirModel {
    * 会自动进行数据别名转换
    * @param jsonArray `JSON`数组
    */
-  static fromJsonArray<T extends AirModel>(this: new () => T, jsonArray: IJson | IJson[] = []): T[] {
+  static fromJsonArray<T extends AirModel>(this: ClassConstructor<T>, jsonArray: IJson | IJson[] = []): T[] {
     const instanceList: T[] = []
     if (Array.isArray(jsonArray)) {
       for (let i = 0; i < jsonArray.length; i += 1) {
@@ -176,10 +177,10 @@ export class AirModel {
     // 最后删除无用的数据
     for (const fieldKey of fieldKeyList) {
       const fieldAliasName = getAlias(instance, fieldKey)
-
-      if (fieldAliasName && fieldAliasName !== fieldKey) {
-        delete (instance as IJson)[fieldAliasName]
+      if (fieldAliasName === fieldKey) {
+        continue
       }
+      delete (instance as IJson)[fieldAliasName]
     }
     return instance
   }
@@ -208,7 +209,7 @@ export class AirModel {
    * @param recoverBy `可选` 初始化用于覆盖对象实例的 `JSON`
    */
   // eslint-disable-next-line no-unused-vars
-  static newInstance<T extends AirModel>(this: new () => T, recoverBy?: IJson): T {
+  static newInstance<T extends AirModel>(this: ClassConstructor<T>, recoverBy?: IJson): T {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     const instance = (Object.assign(new this(), null)) as T
@@ -284,6 +285,10 @@ export class AirModel {
     const json: IJson = {}
     for (const fieldKey of fieldKeyList) {
       const fieldData = (this as IJson)[fieldKey]
+      if (fieldData === null || fieldData === undefined) {
+        // 如果属性值为 null 或 undefined 则不转换到JSON
+        continue
+      }
       let fieldAliasName = getAlias(this, fieldKey) || fieldKey
       if (!getNoPrefix(this, fieldKey) && getFieldPrefix(this)) {
         // 按忽略前缀规则获取别名