+++ config/AirConfig.ts
@@ -22,7 +22,7 @@ export class AirConfig {
   /**
    * ## `AirPower` 版本号
    */
-  static readonly version = 'v2.3.1'
+  static readonly version = 'v2.3.2'
 
   /**
    * ## `AppKey`