+++ helper/AirValidator.ts
@@ -1,8 +1,3 @@
-/* eslint-disable no-case-declarations */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/ban-types */
-/* eslint-disable no-unused-vars */
-
 import { AirInputType } from '../enum/AirInputType'
 import { AirNotification } from '../feedback/AirNotification'
 import { IValidateRule } from '../interface/IValidateRule'
@@ -11,6 +6,7 @@ import { AirEntity } from '../base/AirEntity'
 import { AirAbstractEntityService } from '../base/AirAbstractEntityService'
 import { AirI18n } from './AirI18n'
 import { IJson } from '../interface/IJson'
+import { AirAny, AirValidatorCallback, AirValidatorRule } from '../type/AirType'
 
 /**
  * # 表单验证工具
@@ -50,7 +46,8 @@ export class AirValidator {
    * ## 自定义验证器
    * 请调用 `.setCustomValidator()`
    */
-  private validator!: (rule: any, value: any, callback: Function) => void
+  // eslint-disable-next-line no-unused-vars
+  private validator!: (rule: AirValidatorRule, value: AirAny, callback: AirValidatorCallback) => void
 
   /**
    * ## 通过指定错误信息来创建一个验证器
@@ -157,41 +154,35 @@ export class AirValidator {
     if (str.length !== 18 && str.length !== 15) {
       return false
     }
-    switch (str.length) {
-      case 18:
-        const year = parseInt(str.substring(6), this.DEFAULT_RADIX)
-        if (year > new Date().getFullYear() || year < 1900) {
-          return false
-        }
-        const month = parseInt(str.substring(10, 12), this.DEFAULT_RADIX)
-        if (month > 12 || month < 1) {
-          return false
-        }
-        const day = parseInt(str.substring(12, 14), this.DEFAULT_RADIX)
-        if (day > 31 || month < 1) {
-          return false
-        }
-        const arr: Array<Array<number | 'X'>> = [[7, 9, 10, 5, 8, 4, 2, 1, 6, 3, 7, 9, 10, 5, 8, 4, 2], [1, 0, 'X', 9, 8, 7, 6, 5, 4, 3, 2]]
-        let sum = 0
-        for (let i = 0; i < 17; i += 1) {
-          sum += parseInt(str[i], this.DEFAULT_RADIX) * (arr[0][i] as number)
-        }
-        // eslint-disable-next-line eqeqeq
-        if (arr[1][(sum % 11)] == str[17]) {
-          return true
-        }
-        break
-      case 15:
-        // 15位省份证校验
-        const reg = /^[1-9]\d{5}((\d{2}(((0[13578]|1[02])(0[1-9]|[12][0-9]|3[01]))|((0[13456789]|1[012])(0[1-9]|[12][0-9]|30))|(02(0[1-9]|1[0-9]|2[0-8]))))|(((0[48]|[2468][048]|[13579][26])|(00))0229))\d{2}[0-9Xx]$/
-        if (reg.test(str)) {
-          return true
-        }
-        break
-      default:
+
+    const validArray: Array<Array<number | 'X'>> = [[7, 9, 10, 5, 8, 4, 2, 1, 6, 3, 7, 9, 10, 5, 8, 4, 2], [1, 0, 'X', 9, 8, 7, 6, 5, 4, 3, 2]]
+
+    if (str.length === 15) {
+      // 15位身份证校验
+      return /^[1-9]\d{5}((\d{2}(((0[13578]|1[02])(0[1-9]|[12][0-9]|3[01]))|((0[13456789]|1[012])(0[1-9]|[12][0-9]|30))|(02(0[1-9]|1[0-9]|2[0-8]))))|(((0[48]|[2468][048]|[13579][26])|(00))0229))\d{2}[0-9Xx]$/.test(str)
+    }
+    if (str.length !== 18) {
+      return false
     }
 
-    return false
+    const year = parseInt(str.substring(6), this.DEFAULT_RADIX)
+    if (year > new Date().getFullYear() || year < 1900) {
+      return false
+    }
+    const month = parseInt(str.substring(10, 12), this.DEFAULT_RADIX)
+    if (month > 12 || month < 1) {
+      return false
+    }
+    const day = parseInt(str.substring(12, 14), this.DEFAULT_RADIX)
+    if (day > 31 || month < 1) {
+      return false
+    }
+    let sum = 0
+    for (let i = 0; i < 17; i += 1) {
+      sum += parseInt(str[i], this.DEFAULT_RADIX) * (validArray[0][i] as number)
+    }
+    // eslint-disable-next-line eqeqeq
+    return validArray[1][(sum % 11)] == str[17]
   }
 
   /**
@@ -340,7 +331,7 @@ export class AirValidator {
    * @param str 内容
    */
   ifEquals(str: string): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (value && value === str) {
         callback(this.message || `输入的内容不能是${str}`)
       } else {
@@ -355,7 +346,7 @@ export class AirValidator {
    * @param min 最小值
    */
   ifLengthLessThan(min: number): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (value && value.length < min) {
         callback(this.message || `最少请输入${min}个字符`)
       } else {
@@ -370,7 +361,7 @@ export class AirValidator {
    * @param max 最大值
    */
   ifLengthGreaterThan(max: number): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (value && value.length > max) {
         callback(this.message || `最多允许输入${max}个字符`)
       } else {
@@ -386,7 +377,7 @@ export class AirValidator {
    */
   ifNotLessThan(min: number): this {
     this.toNumber()
-    this.validator = (_: any, value: number, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: number, callback: AirValidatorCallback) => {
       if (value <= min) {
         callback(this.message || `数字必须大于${min}`)
       } else {
@@ -404,7 +395,7 @@ export class AirValidator {
    */
   ifNotGreaterThan(max: number): this {
     this.toNumber()
-    this.validator = (_: any, value: number, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: number, callback: AirValidatorCallback) => {
       if (value >= max) {
         callback(this.message || `数字必须小于${max}`)
       } else {
@@ -420,7 +411,7 @@ export class AirValidator {
    */
   ifLessThan(min: number): this {
     this.toNumber()
-    this.validator = (_: any, value: number, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: number, callback: AirValidatorCallback) => {
       if (value < min) {
         callback(this.message || `数字最小允许输入${min}`)
       } else {
@@ -436,7 +427,7 @@ export class AirValidator {
    */
   ifGreaterThan(max: number): this {
     this.toNumber()
-    this.validator = (_: any, value: number, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: number, callback: AirValidatorCallback) => {
       if (value > max) {
         callback(this.message || `数字最大允许输入${max}`)
       } else {
@@ -469,7 +460,8 @@ export class AirValidator {
    * ## 设置自定义验证器
    * @param validator 验证方法
    */
-  setCustomValidator(validator: (_: any, value: any, callback: Function) => void): this {
+  // eslint-disable-next-line no-unused-vars
+  setCustomValidator(validator: (_: AirAny, value: unknown, callback: AirValidatorCallback) => void): this {
     this.validator = validator
     return this
   }
@@ -480,7 +472,7 @@ export class AirValidator {
    */
   ifNotContain(...whats: string[]): this {
     let error = false
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       error = false
       for (const what of whats) {
         if (!value || value.indexOf(what) < 0) {
@@ -503,7 +495,7 @@ export class AirValidator {
    */
   ifContain(...whats: string[]): this {
     let error = ''
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value) {
         callback()
         return
@@ -529,7 +521,7 @@ export class AirValidator {
    * @param regx 正则
    */
   ifTest(regx: RegExp): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirAny, value: string, callback: AirValidatorCallback) => {
       if (value && regx.test(value)) {
         callback(this.message || AirI18n.get().TestError || '正则表达式校验失败')
       } else {
@@ -544,7 +536,7 @@ export class AirValidator {
    * @param regx 正则
    */
   ifNotTest(regx: RegExp): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (value && !regx.test(value)) {
         callback(this.message || AirI18n.get().TestError || '正则表达式校验失败')
       } else {
@@ -558,7 +550,7 @@ export class AirValidator {
    * ## 如果不是邮箱时报错
    */
   ifNotEmail(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isEmail(value)) {
         callback()
       } else {
@@ -572,7 +564,7 @@ export class AirValidator {
    * ## 如果不是手机号时报错
    */
   ifNotMobilePhone(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isMobilePhone(value)) {
         callback()
       } else {
@@ -586,7 +578,7 @@ export class AirValidator {
    * ## 如果不是座机号时报错
    */
   ifNotTelPhone(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isTelephone(value)) {
         callback()
       } else {
@@ -600,7 +592,7 @@ export class AirValidator {
    * ## 如果不是联系电话时报错
    */
   ifNotPhone(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isTelephoneOrMobilePhone(value)) {
         callback()
       } else {
@@ -614,7 +606,7 @@ export class AirValidator {
    * ## 如果不是纯字母时报错
    */
   ifNotOnlyLetter(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isOnlyLetter(value)) {
         callback()
       } else {
@@ -628,7 +620,7 @@ export class AirValidator {
    * ## 如果不是字母和数字报错
    */
   ifNotOnlyNumberAndLetter(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isOnlyNumberAndLetter(value)) {
         callback()
       } else {
@@ -643,7 +635,7 @@ export class AirValidator {
    */
   ifNotNaturalInteger(): this {
     this.toNumber()
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isNaturalInteger(value)) {
         callback()
       } else {
@@ -658,7 +650,7 @@ export class AirValidator {
    */
   ifNotNaturalNumber(): this {
     this.toNumber()
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isNaturalNumber(value)) {
         callback()
       } else {
@@ -673,7 +665,7 @@ export class AirValidator {
    */
   ifNotInteger(): this {
     this.toNumber()
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isInteger(value)) {
         callback()
       } else {
@@ -688,7 +680,7 @@ export class AirValidator {
    */
   ifNotNumber(): this {
     this.toNumber()
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isNumber(value)) {
         callback()
       } else {
@@ -702,7 +694,7 @@ export class AirValidator {
    * ## 如果不是有效身份证时报错
    */
   ifNotChineseIdCard(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isChineseIdCard(value)) {
         callback()
       } else {
@@ -716,7 +708,7 @@ export class AirValidator {
    * ## 如果不是纯汉字
    */
   ifNotChinese(): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.isChinese(value)) {
         callback()
       } else {
@@ -731,7 +723,7 @@ export class AirValidator {
    * @param list 范围 枚举或字符
    */
   ifNot(...list: AirInputType[] | string[]): this {
-    this.validator = (_: any, value: string, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: string, callback: AirValidatorCallback) => {
       if (!value || AirValidator.validate(value, list as unknown as AirInputType)) {
         callback()
       } else {
@@ -747,7 +739,7 @@ export class AirValidator {
   ifPayloadEmpty(): this {
     this.required = true
     this.trigger = 'change'
-    this.validator = (_: any, value: Object, callback: Function) => {
+    this.validator = (_: AirValidatorRule, value: object, callback: AirValidatorCallback) => {
       if (value) {
         callback()
       } else {