+++ helper/AirVersion.ts
@@ -7,19 +7,23 @@ import { AirAlert } from '../feedback/AirAlert'
  * @author Hamm.cn
  */
 export class AirVersion {
+  /**
+   * ## `Element Plus` 最低版本
+   */
   private static readonly elementPlusRequired = '2.8.0'
 
   /**
    * ## 检查 `Element Plus`
    */
   static checkElementPlus() {
-    if (this.parseVersion(ElementPlus.version) < this.parseVersion(this.elementPlusRequired)) {
-      if (import.meta.env.DEV) {
-        AirAlert.error(`当前版本过低，请升级到 ${this.elementPlusRequired} 或以上版本`, '升级 Element Plus')
-      } else {
-        console.warn(`%cElement Plus 版本过低，请升级至 ${this.elementPlusRequired} 或以上版本`, 'color:red;font-size:12px;')
-      }
+    if (this.parseVersion(ElementPlus.version) >= this.parseVersion(this.elementPlusRequired)) {
+      return
     }
+    if (import.meta.env.DEV) {
+      AirAlert.error(`当前版本过低，请升级到 ${this.elementPlusRequired} 或以上版本`, '升级 Element Plus')
+      return
+    }
+    console.warn(`%cElement Plus 版本过低，请升级至 ${this.elementPlusRequired} 或以上版本`, 'color:red;font-size:12px;')
   }
 
   /**